# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This the total new Krateo backend.",
        "title": "SnowPlow API",
        "contact": {},
        "version": "0.1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/convert": {
            "post": {
                "description": "This endpoint converts YAML to JSON or JSON to YAML based on the \"Content-Type\" header.",
                "consumes": [
                    "application/json",
                    " application/x-yaml",
                    " text/yaml"
                ],
                "produces": [
                    "application/json",
                    " application/x-yaml"
                ],
                "summary": "Convert YAML to JSON or JSON to YAML",
                "operationId": "convert",
                "parameters": [
                    {
                        "description": "Input data in YAML or JSON format",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Converted output in the requested format",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad request, invalid input",
                        "schema": {
                            "$ref": "#/definitions/response.Status"
                        }
                    },
                    "406": {
                        "description": "Unsupported 'Accept' header",
                        "schema": {
                            "$ref": "#/definitions/response.Status"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/response.Status"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "response.Status": {
            "type": "object",
            "properties": {
                "apiVersion": {
                    "type": "string"
                },
                "code": {
                    "description": "Suggested HTTP return code for this status, 0 if not set.",
                    "type": "integer"
                },
                "kind": {
                    "type": "string"
                },
                "message": {
                    "description": "A human-readable description of the status of this operation.",
                    "type": "string"
                },
                "reason": {
                    "description": "A machine-readable description of why this operation is in the\n\"Failure\" status. If this value is empty there\nis no information available. A Reason clarifies an HTTP status\ncode but does not override it.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/response.StatusReason"
                        }
                    ]
                },
                "status": {
                    "description": "Status of the operation.\nOne of: \"Success\" or \"Failure\".",
                    "type": "string"
                }
            }
        },
        "response.StatusReason": {
            "type": "string",
            "enum": [
                "",
                "Unauthorized",
                "Forbidden",
                "NotFound",
                "Conflict",
                "Gone",
                "Invalid",
                "Timeout",
                "TooManyRequests",
                "BadRequest",
                "MethodNotAllowed",
                "NotAcceptable",
                "RequestEntityTooLarge",
                "UnsupportedMediaType",
                "UnprocessableEntity",
                "InternalError",
                "ServiceUnavailable"
            ],
            "x-enum-varnames": [
                "StatusReasonUnknown",
                "StatusReasonUnauthorized",
                "StatusReasonForbidden",
                "StatusReasonNotFound",
                "StatusReasonConflict",
                "StatusReasonGone",
                "StatusReasonInvalid",
                "StatusReasonTimeout",
                "StatusReasonTooManyRequests",
                "StatusReasonBadRequest",
                "StatusReasonMethodNotAllowed",
                "StatusReasonNotAcceptable",
                "StatusReasonRequestEntityTooLarge",
                "StatusReasonUnsupportedMediaType",
                "StatusUnprocessableEntity",
                "StatusReasonInternalError",
                "StatusReasonServiceUnavailable"
            ]
        }
    }
}